+++ mangooio-core/src/main/java/io/mangoo/admin/AdminFilter.java
@@ -1,11 +1,11 @@
 package io.mangoo.admin;
 
+import java.util.Base64;
 import java.util.Objects;
 import java.util.Optional;
 
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.bouncycastle.util.encoders.Base64;
 
 import com.google.common.base.Charsets;
 
@@ -26,6 +26,7 @@ import io.undertow.util.Headers;
  *
  */
 public class AdminFilter implements MangooFilter {
+    private static final Base64.Decoder base64Decoder = Base64.getDecoder();
     private static final Config CONFIG = Application.getConfig();
     
     @Override
@@ -60,7 +61,7 @@ public class AdminFilter implements MangooFilter {
         if (StringUtils.isNotBlank(authInfo)) {
             authInfo = authInfo.replace("Basic", "");
             authInfo = authInfo.trim();
-            authInfo = new String(Base64.decode(authInfo), Charsets.UTF_8);
+            authInfo = new String(base64Decoder.decode(authInfo), Charsets.UTF_8);
 
             final String [] credentials = authInfo.split(":");
             if (credentials != null && credentials.length == Default.BASICAUTH_CREDENTIALS_LENGTH.toInt()) {