+++ mangooio-core/src/main/java/io/mangoo/filters/BasicAuthenticationFilter.java
@@ -1,7 +1,8 @@
 package io.mangoo.filters;
 
+import java.util.Base64;
+
 import org.apache.commons.lang3.StringUtils;
-import org.bouncycastle.util.encoders.Base64;
 
 import com.google.common.base.Charsets;
 
@@ -22,6 +23,7 @@ import io.undertow.util.Headers;
  *
  */
 public class BasicAuthenticationFilter implements MangooFilter {
+    private static final Base64.Decoder base64Decoder = Base64.getDecoder();
     private static final Config CONFIG = Application.getConfig();
     
     @Override
@@ -32,7 +34,7 @@ public class BasicAuthenticationFilter implements MangooFilter {
         if (StringUtils.isNotBlank(authInfo)) {
             authInfo = authInfo.replace("Basic", "");
             authInfo = authInfo.trim();
-            authInfo = new String(Base64.decode(authInfo), Charsets.UTF_8);
+            authInfo = new String(base64Decoder.decode(authInfo), Charsets.UTF_8);
 
             String [] credentials = authInfo.split(":");
             if (credentials != null && credentials.length == Default.BASICAUTH_CREDENTIALS_LENGTH.toInt()) {